memperbaiki class main
