+++ nbproject/private/private.xml
@@ -4,6 +4,7 @@
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
             <file>file:/D:/Kuliah/PBO/PROJEK%20java/TUBES_KELOMPOK%203_APK_GO-LEARNLISH/src/tubes_kelompok/pkg3_apk_go/learnlish/NewJFrame.java</file>
+            <file>file:/D:/Kuliah/PBO/PROJEK%20java/TUBES_KELOMPOK%203_APK_GO-LEARNLISH/src/tubes_kelompok/pkg3_apk_go/learnlish/kuis.java</file>
             <file>file:/D:/Kuliah/PBO/PROJEK%20java/TUBES_KELOMPOK%203_APK_GO-LEARNLISH/src/tubes_kelompok/pkg3_apk_go/learnlish/Form_login.java</file>
             <file>file:/D:/Kuliah/PBO/PROJEK%20java/TUBES_KELOMPOK%203_APK_GO-LEARNLISH/src/tubes_kelompok/pkg3_apk_go/learnlish/NewPage.java</file>
         </group>
