+++ nbproject/project.properties
@@ -54,7 +54,7 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=tubes_kelompok.pkg3_apk_go.learnlish.LoginFormDemo
+main.class=tubes_kelompok.pkg3_apk_go.learnlish.mainlogin
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
