+++ src/tubes_kelompok/pkg3_apk_go/learnlish/NewPage.java
@@ -1,4 +1,5 @@
-//import required classes and packages  
+//import required classes and packages 
+package tubes_kelompok.pkg3_apk_go.learnlish;
 import javax.swing.*;  
 import java.awt.*;
